server/requirements.txt
django>=4.2.7, <4.3
djangorestframework>=3.14.0,<3.15
django-channels>=0.7.0, <0.8
channels>=4.0.0, <4.1
psycopg2>=2.9.9, <2.10
Pillow>=10.1.0, <10.2
pandas>=2.1.3, <2.2
numpy>=1.26.2, <1.27
